+++ CHANGELOG.md
@@ -41,6 +41,11 @@
     * Renamed and replaced `BTPayPalNativeCheckoutClient.tokenizePayPalAccount` with two methods called `tokenize()` taking in requests of either `BTPayPalNativeCheckoutRequest` or `BTPayPalNativeVaultRequest`
     * `BTPayPalNativeCheckoutRequest` now takes in an `intent` of type `BTPayPalRequestIntent` instead of `BTPayPalNativeRequestIntent`
     * `BTPayPalNativeCheckoutRequest.localeCode` now uses the `BTPayPalLocaleCode` enum instead of a `String` value
+  * BraintreeUnionPay
+    * Remove `BraintreeUnionPay` module
+      * UnionPay cards can now be processed as regular cards (through the BraintreeCard module) due to their partnership with Discover
+  * BraintreeCore
+    * Remove `BTConfiguration+Extensions.isUnionPayEnabled` property
 
 ## 5.19.0 (2022-12-19)
 * BraintreePayPalNativeCheckout (BETA)